binance-futures-trading-bot/trading_bot/requirements.txt
python-binance>=1.0.20,<2.0.0
python-dotenv>=1.0.1,<2.0.0
colorama>=0.4.6,<1.0.0
